+++ src/main/java/lermitage/intellij/extra/icons/ExtraIconPatcher.java
@@ -5,14 +5,22 @@ package lermitage.intellij.extra.icons;
 import com.intellij.openapi.diagnostic.Logger;
 import com.intellij.openapi.util.IconLoader;
 import com.intellij.openapi.util.IconPathPatcher;
-import lermitage.intellij.extra.icons.cfg.SettingsService;
+import com.intellij.util.Base64;
+import lermitage.intellij.extra.icons.cfg.services.impl.SettingsIDEService;
+import org.apache.commons.io.FileUtils;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import java.io.File;
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.util.Collection;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 public class ExtraIconPatcher extends IconPathPatcher {
 
@@ -20,15 +28,16 @@ public class ExtraIconPatcher extends IconPathPatcher {
 
     private Map<String, String> icons;
 
+
     @NotNull
     public static Map<String, String> getEnabledIcons() {
         Map<String, String> enabledIcons = new LinkedHashMap<>();
-        List<String> disabledModelIds = SettingsService.getIDEInstance().getDisabledModelIds();
-        ExtraIconProvider.allModels()
-            .stream()
+        List<String> disabledModelIds = SettingsIDEService.getIDEInstance().getDisabledModelIds();
+        Stream.of(ExtraIconProvider.allModels(), SettingsIDEService.getInstance().getCustomModels()).flatMap(Collection::stream)
+            .collect(Collectors.toList()).stream()
             .filter(model -> model.getModelType() == ModelType.ICON)
             .forEach(model -> {
-                if (!enabledIcons.containsKey(model.getIdeIcon()) && !disabledModelIds.contains(model.getId())) {
+                if (model.isEnabled() && !enabledIcons.containsKey(model.getIdeIcon()) && !disabledModelIds.contains(model.getId())) {
                     enabledIcons.put(model.getIdeIcon(), model.getIcon());
                 }
             });
@@ -51,12 +60,43 @@ public class ExtraIconPatcher extends IconPathPatcher {
         String iconOriginalPath = (new File(path)).getName();
         if (icons == null) {
             loadConfig();
+
         }
         return this.icons.get(iconOriginalPath);
     }
 
     private void loadConfig() {
         icons = getEnabledIcons();
+        try {
+            icons = convertB64IconsToLocalFiles(icons);
+        } catch (IOException e) {
+            LOG.warn("Cannot create temporary directory to store user IDE icons, this feature won't work", e);
+        }
+        //icons.put("add.svg", "file://C:\\Projects\\ij-extra-icons\\src\\main\\resources\\extra-icons\\angular2.svg");
         LOG.info("config loaded with success, enabled " + icons.size() + " items");
+    }
+
+    /**
+     * Convert Base64 user icons to local files. This conversion is needed because IconPatcher can't work with
+     * in-memory byte arrays; we have to use bundled icons (from /resources) or local files. This method
+     * stores in-memory Base64 icons provided by user as temporary local files.
+     */
+    private Map<String, String> convertB64IconsToLocalFiles(Map<String, String> icons) throws IOException {
+        Map<String, String> morphedIcons = new LinkedHashMap<>();
+        for (String iconKey : icons.keySet()) {
+            String iconStr = icons.get(iconKey);
+            if (iconStr.startsWith("/extra-icons/")) {
+                // bundled icon, nothing to do
+                morphedIcons.put(iconKey, iconStr);
+            } else {
+                // base64 icon provided by user: store as local file
+                Path iconFile = Files.createTempFile("extra-icons-ide-user-icon", ".svg");
+                FileUtils.forceDeleteOnExit(iconFile.toFile());
+                Files.write(iconFile, Base64.decode(iconStr));
+                String decodedIconPath = iconFile.toAbsolutePath().toString();
+                morphedIcons.put(iconKey, "file://" + decodedIconPath);
+            }
+        }
+        return morphedIcons;
     }
 }