+++ src/main/java/lermitage/intellij/extra/icons/cfg/dialogs/ModelDialog.java
@@ -39,7 +39,12 @@ import javax.swing.ListCellRenderer;
 import javax.swing.SwingUtilities;
 import java.awt.BorderLayout;
 import java.awt.Component;
+import java.awt.Cursor;
+import java.awt.Desktop;
 import java.awt.event.ItemEvent;
+import java.awt.event.MouseAdapter;
+import java.awt.event.MouseEvent;
+import java.net.URI;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
@@ -67,6 +72,9 @@ public class ModelDialog extends DialogWrapper {
     private JPanel conditionsPanel;
     private JBLabel idLabel;
     private JComboBox<Object> chooseIconSelector;
+    private JBTextField ideIconOverrideTextField;
+    private JLabel ideIconOverrideLabel;
+    private JBLabel ideIconOverrideTip;
 
     private CustomIconLoader.ImageWrapper customIconImage;
     private JPanel toolbarPanel;
@@ -89,6 +97,20 @@ public class ModelDialog extends DialogWrapper {
 
     private void initComponents() {
         setIdComponentsVisible(false);
+        ideIconOverrideTip.setText("<html><b>Find IDE icons <a href=\"#\">here</a></b>, open the ZIP file then use " +
+            "the non-dark icon name (with <i>.svg</i> extension).</html>");
+        ideIconOverrideTip.setToolTipText("<html>Open <i>https://jetbrains.design/intellij/resources/icons_list/</i> in default browser.</html>");
+        ideIconOverrideTip.setCursor(Cursor.getPredefinedCursor(Cursor.HAND_CURSOR));
+        ideIconOverrideTip.addMouseListener(new MouseAdapter() {
+            @Override
+            public void mouseClicked(MouseEvent e) {
+                try {
+                    Desktop.getDesktop().browse(new URI("https://jetbrains.design/intellij/resources/icons_list/"));
+                } catch (Exception ex) {
+                    //ignore
+                }
+            }
+        });
         conditionsCheckboxList = new CheckBoxList<>((index, value) -> {
             //noinspection ConstantConditions
             conditionsCheckboxList.getItemAt(index).setEnabled(value);
@@ -112,6 +134,11 @@ public class ModelDialog extends DialogWrapper {
 
         typeComboBox.addItem(ModelType.FILE.getFriendlyName());
         typeComboBox.addItem(ModelType.DIR.getFriendlyName());
+        if (!settingsForm.isProjectForm()) {
+            typeComboBox.addItem(ModelType.ICON.getFriendlyName());
+        }
+
+        typeComboBox.addActionListener(e -> updateUIOnTypeChange());
 
         chooseIconSelector.addItem("choose custom or bundled icon");
         List<Model> bundledModels = new ArrayList<>();
@@ -139,6 +166,19 @@ public class ModelDialog extends DialogWrapper {
                 }
             }
         });
+
+        updateUIOnTypeChange();
+    }
+
+    private void updateUIOnTypeChange() {
+        Object selectedItem = typeComboBox.getSelectedItem();
+        if (selectedItem != null) {
+            boolean ideIconOverrideSelected = selectedItem.equals(ModelType.ICON.getFriendlyName());
+            ideIconOverrideLabel.setVisible(ideIconOverrideSelected);
+            ideIconOverrideTextField.setVisible(ideIconOverrideSelected);
+            ideIconOverrideTip.setVisible(ideIconOverrideSelected);
+            conditionsPanel.setVisible(!ideIconOverrideSelected);
+        }
     }
 
     /**
@@ -158,15 +198,29 @@ public class ModelDialog extends DialogWrapper {
             icon = modelToEdit.getIcon();
             iconType = modelToEdit.getIconType();
         }
-        Model newModel = new Model(modelIDField.isVisible() ? modelIDField.getText() : null,
-            icon,
-            descriptionField.getText(),
-            ModelType.getByFriendlyName(Objects.requireNonNull(typeComboBox.getSelectedItem()).toString()),
-            iconType,
-            IntStream.range(0, conditionsCheckboxList.getItemsCount())
-                .mapToObj(index -> conditionsCheckboxList.getItemAt(index))
-                .collect(Collectors.toList())
-        );
+
+        Object selectedItem = typeComboBox.getSelectedItem();
+        Model newModel;
+        if (selectedItem != null && selectedItem.equals(ModelType.ICON.getFriendlyName())) {
+            newModel = new Model(modelIDField.isVisible() ? modelIDField.getText() : null,
+                ideIconOverrideTextField.getText(),
+                icon,
+                descriptionField.getText(),
+                ModelType.getByFriendlyName(Objects.requireNonNull(typeComboBox.getSelectedItem()).toString()),
+                iconType
+            );
+        } else {
+            newModel = new Model(modelIDField.isVisible() ? modelIDField.getText() : null,
+                icon,
+                descriptionField.getText(),
+                ModelType.getByFriendlyName(Objects.requireNonNull(typeComboBox.getSelectedItem()).toString()),
+                iconType,
+                IntStream.range(0, conditionsCheckboxList.getItemsCount())
+                    .mapToObj(index -> conditionsCheckboxList.getItemAt(index))
+                    .collect(Collectors.toList())
+            );
+        }
+
         if (modelToEdit != null) {
             newModel.setEnabled(modelToEdit.isEnabled());
         }
@@ -185,6 +239,7 @@ public class ModelDialog extends DialogWrapper {
             modelIDField.setText(model.getId());
         }
         descriptionField.setText(model.getDescription());
+        ideIconOverrideTextField.setText(model.getIdeIcon());
         typeComboBox.setSelectedItem(model.getModelType().getFriendlyName());
         typeComboBox.updateUI();
         Double additionalUIScale = SettingsService.getIDEInstance().getAdditionalUIScale();
@@ -200,6 +255,8 @@ public class ModelDialog extends DialogWrapper {
         }
         model.getConditions().forEach(modelCondition ->
             conditionsCheckboxList.addItem(modelCondition, modelCondition.asReadableString(FIELD_SEPARATOR), modelCondition.isEnabled()));
+
+        updateUIOnTypeChange();
     }
 
     /**
@@ -261,9 +318,18 @@ public class ModelDialog extends DialogWrapper {
         if (customIconImage == null && modelToEdit == null) {
             return new ValidationInfo("Please add an icon!", chooseIconButton);
         }
-        if (conditionsCheckboxList.isEmpty()) {
-            return new ValidationInfo("Please add a condition to your model!", toolbarPanel);
+
+        Object selectedItem = typeComboBox.getSelectedItem();
+        if (selectedItem != null && selectedItem.equals(ModelType.ICON.getFriendlyName())) {
+            if (ideIconOverrideTextField.getText().trim().isEmpty()) {
+                return new ValidationInfo("IDE icon's name cannot be empty!", ideIconOverrideTextField);
+            }
+        } else {
+            if (conditionsCheckboxList.isEmpty()) {
+                return new ValidationInfo("Please add a condition to your model!", toolbarPanel);
+            }
         }
+
         return super.doValidate();
     }